Rename library name
