+++ package.json
@@ -1,12 +1,12 @@
 {
-  "name": "simple-redux-persist",
+  "name": "painless-redux-persist",
   "version": "1.0.0",
   "description": "Synchronize Redux state with localStorage or sessionStorage",
   "license": "AGPL-3.0-or-later",
   "author": "Christian Llontop",
-  "main": "dist/simple-redux-persist.cjs.js",
-  "module": "dist/simple-redux-persist.esm.js",
-  "browser": "dist/simple-redux-persist.umd.js",
+  "main": "dist/painless-redux-persist.cjs.js",
+  "module": "dist/painless-redux-persist.esm.js",
+  "browser": "dist/painless-redux-persist.umd.js",
   "scripts": {
     "prepare": "yarn build",
     "build": "rollup -c"
@@ -27,7 +27,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/christianllv/simple-redux-persist.git"
+    "url": "https://github.com/christianllv/painless-redux-persist.git"
   },
   "keywords": [
     "redux",
@@ -37,9 +37,9 @@
     "state"
   ],
   "bugs": {
-    "url": "https://github.com/christianllv/simple-redux-persist/issues"
+    "url": "https://github.com/christianllv/painless-redux-persist/issues"
   },
-  "homepage": "hhttps://github.com/christianllv/simple-redux-persist",
+  "homepage": "hhttps://github.com/christianllv/painless-redux-persist",
   "files": [
     "dist"
   ]
