+++ src/index.js
@@ -1,3 +1,3 @@
 export {
   default, defineState, getState, resetState, storeConfig,
-} from './simple-redux-persist';
+} from './painless-redux-persist';
