+++ src/painless-redux-persist.js
